backend/requirements.txt
# Supply Chain Backend — pinned requirements
# Generated from pip freeze on 2026-04-28

fastapi==0.136.0
uvicorn[standard]==0.44.0
python-dotenv==1.2.2
groq==1.2.0
pydantic[email]==2.13.3
pydantic-settings==2.14.0
email-validator==2.3.0
python-multipart==0.0.27

# Database (PostgreSQL for cloud, SQLite fallback for local)
sqlalchemy==2.0.49
psycopg2-binary==2.9.11

# Auth
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP / networking
httpx==0.28.1
requests==2.33.1
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
urllib3==2.6.3

# Testing
pytest==9.0.3
pytest-anyio==0.0.0
anyio==4.13.0

# Utilities
starlette==1.0.0
annotated-types==0.7.0
typing_extensions==4.15.0
sniffio==1.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
watchfiles==1.1.1
websockets==16.0
click==8.3.2
google-generativeai>=0.8.0
